SCRD-2067 fix empty bind_host error in Ocata
When thet [DEFAULT]/bind_host variable is set to an empty value,
the neutron services fail shortly after startup with the following
error:

  Unhandled error: ConfigFileValueError: Value for option bind_host
  is not valid:  is not a valid host address

, even if this option is only actually used by neutron-server.

On nodes other than the neutron server nodes (e.g. nova-compute
and neutron-network nodes), the host.bind.NEU_SVR variable is
not defined, which is causing these problems. To avoid this
error the variable is not added to the configuration file,
defaulting to the neutron default value (currently 0.0.0.0).

Change-Id: I02e63b971583e4ae6b486b5c29e25d9e165bd94d