SCRD-999 Set default location for vhostuser sockets
When binding DPDK ports, the OVS agent will communicate this
directory to Nova, which in turn will tell QEMU to start
listnening on a Unix socket created in such directory. QEMU
server mode [1] is the only sane way to integrate QEMU/DPDK
and to preserve data plane connectivity across OVS restarts.

For this reason, it's wise to add /var/lib/libvirt/qemu as
sane default in ML2/OVS.

[1] https://software.intel.com/en-us/articles/data-plane-development-kit-vhost-user-client-mode-with-open-vswitch

Change-Id: If8cd8cb744749c24c5800ebf6ce167c869b7e362