SCRD-5359 Neutron vpn-agent becomes l3 agent extension
The neutron-vpn-agent has been converted to be an extension of the
neutron-l3-agent in Queens.

As a first pass to implementing this we change the executable that
we run the provide the neutron-vpn-agent service, but leave the rest
of the neutron-vpn-agent setup the same.

We also change the l3_agent.ini templating to generate the list of
extensions based upon an ansible fact that we dynamically setup based
upon what services are defined for the group the play is running on,
defaulting to just the fwaas like it was before.

And we add the vpnagent section to the l3_agent.ini file if the vpnaas
extension is being enabled.

Change-Id: I79a6da5c576124557f886d19a914f76b43c654de